[WIP] Exploring 64bits source location

+++ clang/include/clang-c/CXSourceLocation.h
@@ -43,7 +43,7 @@ LLVM_CLANG_C_EXTERN_C_BEGIN
  */
 typedef struct {
   const void *ptr_data[2];
-  unsigned int_data;
+  unsigned long long int_data;
 } CXSourceLocation;
 
 /**
@@ -54,8 +54,8 @@ typedef struct {
  */
 typedef struct {
   const void *ptr_data[2];
-  unsigned begin_int_data;
-  unsigned end_int_data;
+  unsigned long long begin_int_data;
+  unsigned long long end_int_data;
 } CXSourceRange;
 
 /**
